TelegramNotif/requirements.txt
aiohttp    ==         3.9.1
aiosignal  ==         1.3.1
anyio      ==         4.2.0
APScheduler     ==    3.10.4
async-timeout    ==   4.0.3
attrs      ==         23.1.0
certifi    ==         2023.11.17
charset-normalizer == 3.3.2
confluent-kafka  ==   2.3.0
exceptiongroup  ==    1.2.0
frozenlist ==         1.4.1
h11        ==         0.14.0
httpcore   ==         1.0.2
httpx      ==         0.25.2
idna       ==         3.6
logger     ==         1.4
multidict  ==         6.0.4
pip        ==         23.3.2
python-telegram-bot == 20.7
pytz       ==         2023.3.post1
requests   ==         2.31.0
setuptools ==         68.2.0
six        ==         1.16.0
sniffio    ==         1.3.0
telepot    ==         12.7
typing_extensions  == 4.9.0
tzlocal    ==         5.2
urllib3    ==         2.1.0
wheel      ==         0.41.2
yarl       ==         1.9.4
